feat: creamos contexto y migramos las peticiones
style: seccion portfolio esta iniciada
