+++ portfolio-metamorfose/src/components/ButtonContact/index.tsx
@@ -1,0 +1,7 @@
+export const ButtonContact = () => {
+  return (
+    <button className="absolute font-[family-name:var(--font-bellota)] tracking-widest w-45 h-15 sm:w-50 sm:h-12 lg:w-45 lg:h-12 xl:w-50 xl:h-12 2xl:w-50  2xl:h-12  mt-[70%]  p-2  rounded bg-white text-black hover:text-white transition-all ease-in-out duration-600 font-extrabold hover:scale-102 hover:font-bold hover:cursor-pointer z-50 bg-swipe-hover sm:mt-[6%] masked-reveal">
+      Quiero una sesión
+    </button>
+  );
+};
