+++ portfolio-metamorfose/src/components/CategorySection/index.tsx
@@ -1,0 +1,17 @@
+import { ICategory } from "@/types/types";
+
+export const CategoryComponent: React.FC<{ categories: ICategory[] }> = ({
+  categories,
+}) => {
+  return (
+    <div>
+      {categories.map((category) => {
+        return (
+          <div key={category.id}>
+            <p>{category.name}</p>
+          </div>
+        );
+      })}
+    </div>
+  );
+};
